web/requirements.txt
Flask==3.1.2
google-generativeai>=0.8.0
requests==2.32.4
python-dotenv==1.0.1
xlsxwriter==3.1.9
openpyxl==3.1.2
reportlab==4.0.9
Pillow==10.4.0
PyPDF2==3.0.1
pdfplumber==0.10.4